refactor: modularization functions
