+++ packages/oraidex-sync/src/sync-cmd.ts
@@ -1,0 +1,10 @@
+import { OraiDexSync } from "./index";
+import { DuckDb } from "./db";
+
+async function initSync() {
+  const duckDb = await DuckDb.create("oraidex-sync-data-only-v2");
+  const oraidexSync = await OraiDexSync.create(duckDb, process.env.RPC_URL, process.env as any);
+  oraidexSync.sync();
+}
+
+initSync();
